Velocity PD controller stubs

+++ include/encoder.h
@@ -14,11 +14,11 @@
 void encoderStart(void);
 long getLeftEncoderValue(void);
 long getRightEncoderValue(void);
-void setLeftEncoderValue(long value);
-void setRightEncoderValue(long value);
+void setLeftEncoderValue(double value);
+void setRightEncoderValue(double value);
 
 // left and right encoder values
-extern long leftEncoderValue;
-extern long rightEncoderValue;
+extern double leftEncoderValue;
+extern double rightEncoderValue;
 
 #endif
